+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/createcompany.go
@@ -12,6 +12,8 @@ type CreateCompanyResponse struct {
 	Company *shared.Company
 	// HTTP response content type for this operation
 	ContentType string
+	// The request made is not valid.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -30,6 +32,13 @@ func (o *CreateCompanyResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
+}
+
+func (o *CreateCompanyResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *CreateCompanyResponse) GetStatusCode() int {