+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/setconfiguration.go
@@ -24,6 +24,8 @@ type SetConfigurationResponse struct {
 	Configuration *shared.Configuration
 	// HTTP response content type for this operation
 	ContentType string
+	// The request made is not valid.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -42,6 +44,13 @@ func (o *SetConfigurationResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
+}
+
+func (o *SetConfigurationResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *SetConfigurationResponse) GetStatusCode() int {