+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/tracking.go
@@ -2,21 +2,21 @@
 
 package shared
 
-type TrackingProjectReference struct {
+type TrackingAccountingProjectReference struct {
 	// Unique identifier to the project reference.
 	ID string `json:"id"`
 	// The project's name.
 	Name *string `json:"name,omitempty"`
 }
 
-func (o *TrackingProjectReference) GetID() string {
+func (o *TrackingAccountingProjectReference) GetID() string {
 	if o == nil {
 		return ""
 	}
 	return o.ID
 }
 
-func (o *TrackingProjectReference) GetName() *string {
+func (o *TrackingAccountingProjectReference) GetName() *string {
 	if o == nil {
 		return nil
 	}
@@ -30,8 +30,8 @@ type Tracking struct {
 	// Defines if the bill or bill credit note is billed/rebilled to a project.
 	IsBilledTo BilledToType `json:"isBilledTo"`
 	// Defines if the bill or bill credit note is billed/rebilled to a project.
-	IsRebilledTo BilledToType              `json:"isRebilledTo"`
-	ProjectRef   *TrackingProjectReference `json:"projectRef,omitempty"`
+	IsRebilledTo BilledToType                        `json:"isRebilledTo"`
+	ProjectRef   *TrackingAccountingProjectReference `json:"projectRef,omitempty"`
 	// Links the current record to the underlying record or data type that created it.
 	//
 	// For example, if a journal entry is generated based on an invoice, this property allows you to connect the journal entry to the underlying invoice in our data model.
@@ -66,7 +66,7 @@ func (o *Tracking) GetIsRebilledTo() BilledToType {
 	return o.IsRebilledTo
 }
 
-func (o *Tracking) GetProjectRef() *TrackingProjectReference {
+func (o *Tracking) GetProjectRef() *TrackingAccountingProjectReference {
 	if o == nil {
 		return nil
 	}