+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [go v0.22.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [Go v0.22.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.22.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-10-30 01:24:36
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.22.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [Go v0.22.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.22.1 - previous-versions/sync-for-commerce-version-1